feat: Add build site hook on record unpublish or delete
